+++ RestaurantApi/src/main/java/com/yp/builder/MediaDtoBuilder.java
@@ -6,18 +6,18 @@ import com.yp.entity.Media;
 import static org.apache.tomcat.util.buf.HexUtils.fromHexString;
 
 public class MediaDtoBuilder extends Builder{
-    private int _id;
+    private Long _id;
     private String _name;
     private byte[] _content;
 
     public MediaDtoBuilder(){
-        this._id=1;
+        this._id=1L;
         this._name="";
         this._content= fromHexString("00");
     }
 
     @Override
-    public MediaDtoBuilder withId(int id) {
+    public MediaDtoBuilder withId(Long id) {
         this._id = id;
         return this;
     }
