+++ RestaurantApi/src/main/java/com/yp/builder/UserBuilder.java
@@ -8,14 +8,14 @@ import java.util.Set;
 
 public class UserBuilder extends Builder{
 
-    private int _id;
+    private Long _id;
     private String _userName;
     private String _passWord;
     private boolean _enabled;
     private Set<Authority> _authorities;
 
     public UserBuilder(){
-        this._id=1;
+        this._id=1L;
         this._userName="";
         this._passWord="";
         this._enabled=true;
@@ -24,7 +24,7 @@ public class UserBuilder extends Builder{
     }
 
     @Override
-    public UserBuilder withId(int id) {
+    public UserBuilder withId(Long id) {
         this._id=id;
         return this;
     }
