+++ RestaurantApi/src/main/java/com/yp/repos/SellOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SellOrderRepository extends JpaRepository<SellOrder, Integer> {
+public interface SellOrderRepository extends JpaRepository<SellOrder, Long> {
 }
