+++ RestaurantApi/src/main/java/com/yp/repos/TableCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TableCategoryRepository extends JpaRepository<TableCategory, Integer> {
+public interface TableCategoryRepository extends JpaRepository<TableCategory, Long> {
 }
